Rework where & when files are opened
The truncate open flag is incompatible with read modes because truncate
wipes the file on-open. So now the functions that explicitly access the
file will open the file with their respective modes, do what they need
to do, and then close the file.

I need to write to truncate lol