Move variables & members that can be static to be static
For the variables, this saves uncecessary memory usage, as well as
enabled the static functions to be able to use them. And the static
functions are the only ones that use them. As for the functions, they
don't modify the objects in any way and don't really need to be tied to
an existing object in order to be called.